+++ source3/smbd/service.c
@@ -973,7 +973,7 @@ connection_struct *make_connection_smb2(struct smbd_server_connection *sconn,
 
 connection_struct *make_connection(struct smbd_server_connection *sconn,
 				   const char *service_in,
-				   const char *pdev, uint16 vuid,
+				   const char *pdev, uint64_t vuid,
 				   NTSTATUS *status)
 {
 	uid_t euid;
